pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foundry-pipeline-assistant"
version = "0.1.0"
description = "A pipeline assistant for Foundry data workflows"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-dotenv",
    "httpx",
    "pydantic",
    "click",
    "typing-extensions",
]

[project.scripts]
foundry-pipeline-assistant = "foundry_pipeline_assistant:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "requests-mock",
    "ruff",
]

[project.urls]
Homepage = "https://github.com/yourusername/foundry-pipeline-assistant"
Repository = "https://github.com/yourusername/foundry-pipeline-assistant.git"
Issues = "https://github.com/yourusername/foundry-pipeline-assistant/issues"

[tool.hatch.build.targets.wheel]
packages = ["foundry_pipeline_assistant"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"